examples/integration/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
PyYAML>=6.0.1
python-dotenv>=1.0.0
gitpython>=3.1.40
pydantic>=2.0.0
